Rearrange graph_links.rs
- Rename `GraphLinksConverter` to `GraphLinksSerializer`.
- Extract some structs from `graph_links.rs` into additional files:
  - `graph_links/header.rs`: `HeaderPlain` and `HeaderCompressed`.
  - `graph_links/serializer.rs`: for `GraphLinksSerializer`.
  - `graph_links/view.rs`: `GraphLinksView`.
- In `header.rs`, remove `HeaderCompressed::version` doc comment about
  `HEADER_VERSION_COMPRESSED` as being outdated after merging #5749.
- Add doc comment to `GraphLinksView`.
- In `view.rs`/`GraphLinksView`, move some associated functions to the
  module level:
  - `read_level_offsets`
  - `get_slice`
  - `error_unsufficent_size`
